+++ src/backend/worksheet/plots/cartesian/CartesianPlotPrivate.h
@@ -53,7 +53,7 @@ class CartesianPlotPrivate:public AbstractPlotPrivate{
 
 		virtual void retransform();
 		void retransformScales();
-		float round(float value, int precision);
+		double round(double value, int precision);
 		void checkXRange();
 		void checkYRange();
 		CartesianCoordinateSystem::Scale* createScale(CartesianPlot::Scale type, Interval<double>& interval,
